Changed type of direction in Raycast from double to float, because raycasts are typically not THAT precise anyway, and it's more consistent with RenderView.mouseDirection, which is usually the value for it

+++ OpenXR/src/me/anno/openxr/ecs/VRHandController.kt
@@ -66,7 +66,7 @@ class VRHandController : Component(), OnUpdate {
 
         val query = RayQuery(
             controller.position,
-            controller.rotation.transform(Vector3d(0.0, -1.0, -0.4).normalize()),
+            controller.rotation.transform(Vector3f(0f, -1f, -0.4f).normalize()),
             1e9
         )
